requirements.txt
streamlit
supabase
gspread
google-auth
facenet-pytorch
torch
torchvision
Pillow
numpy